Fix the VC issue ive had for months
So, turns out, the whole hanging thing ("channel.connect is hanging? I cant fix this") was caused by a lack of intents. This whole time, the bot couldnt check if it successfully joined the vc because it couldnt see voice states. Instead of telling me that, it hangs. Oh well!